Enable TooManyFunctions Detekt rule
